+++ __test__/key.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from 'vitest';
-import Key from '../src/key';
-import { type Scale } from '../src/types';
+import Key from '../src/class/key';
+import { type Scale } from '../src/types/types';
 
 // ========== Key ==========
 describe('Key class', () => {
