+++ src/class/error.ts
