+++ src/class/score.ts
